Implement Simple RabbitMQ Message Passing
A simple implementation of rabbitMQ message passing is implemented. This
utilises a rabbit queue/server, specified within .rabbit_config, and
passes a simple message consisting of transaction ID and the http
request contents to a single queue. Obviously this is a simple prototype
and so will invariably be made more complex, but this provides a base to
build off of.

The connection to the rabbit queue is maintained as a global variable
and monitored at message passing time, but this may not be robust
enough. The message passing is only implemented for the methods that
presently exist i.e. get, getlist and put for files.