Add Exchanges and Generalise Config Parsing
The rabbitMQ connection is made slightly more robust through the
addition of a single, 'direct' exchange which will route messages to
different queues based on a routing_key provided at message time. The
routing keys and queue structure have yet to be finalised.

Also added rabbit config to the .server_config file and expanded the
config parsing to be general based off of a schema, now residing in
utils.constants.