Add Rabbit Consumer and Queue-specific Implem.
A base rabbit consumer class and more queue-specific consumer classes
are implemented to provide a prototype of the rabbit message passing
system intended for use in nlds. Partially addresses issue #5 but
requires more specific work in the index consumer to fully do so.

Slight restructuring included, with the rabbit base classes moved to
submodule nlds.rabbit and the worker processes (consumers) included in a
separate module nlds_processors.