Adds Per-consumer Logging to Separate Files
The logging of messages to separate files is added, with the log level
specified as the final part of the routing key (e.g. nlds.log.info) and
the log target/message specified as additional entries within the DATA
section of the message body.

Also adds unit tests and fixes the fixtures in conftest used for
mocking the body and method parameters in consumer callbacks.