Fix file logging and unite consumer/publisher logs
The problem with file output not being written for the nlds logging
consumer is fixed by changing the logging level of the higher level
Logger object associated with each consumer. Additional tweaks are made
to the console logging system, including unification of the consumer and
publisher Logger objects into a single "nlds.root" logger. This allows
the startup message to be properly logged and all manual calls of
log.info etc in the callback functions to work properly.