Updates indexer to check permissions and file size
The indexer is overhauled to improve functionality, with the ability to
check file permissions and file size added (via a call to os.stat) and
more robust failure checking included. Does not now yield as the
requirement to send messages from within the index function emerged. Now
sends three different filelists to the exchange with different success
levels in the routing key (indexed, retry, failed) whereby each file is
attempted to be accessed a maximum of 4 times before being classed as
'failed'.

Might require some further pruning and testing. Addresses \#16, \#20 & \#22