Implement new stat permissions checking in Indexer
The new check_permissions function, built upon checking file permissions
bitmasks is implemented in the indexer consumer. This involved adding
the uid and gid as member variables to be populated (and reset) every
callback. A bug with attempting to increment the immutable retries in
the namedtuple was also fixed.

This is done to allow the consumer to be run not as root, and minimises
the number of stat calls done per file (limited to at most 1 per file or
directory).