Add retry- and failed- lists to Transfer consumer
Added the retry-list and failed-list logic from the indexer into the
transfer processor. It similarly checks for number of retries for the
failedlist, and checks for access (either existence or permissions) for
the retrylist increment. This logic was directly ported from the indexer
by refactoring a few functions back into the base consumer.

Also edited some lines down to the 80 character limit for PEP-8
compliance.