Implement rotating log files
As per issue #29, implemented the TimedRotatingFileHandler for logging,
allowing the user to configure a rotation rollover (one of 'S', 'M',
'H', 'D', or 'W0..6') as per the 'when' kwarg. Default is 'W0', i.e.
weekly starting on a Monday. The interval is currently not configurable
and defaulted to 1, so setting 'D' would be rotating every 1 days.