Fix logger tests following updates
The logger tests are (finally) updated to match the current form of the
logger and log message format.

Notable is the workaround for a problem with pytest's caplog fixture,
which doesn't seem to be able to pick up debug messages. This is a known
issue with no current fix, and will need to be revisited when a fix has
been found.