Add Support for Delayed Messaging
The delayed messaging rabbitmq plugin is now directly supported within
nlds. Multiple exchanges can be declared and each can be specified as a
'delayed' exchange, whereby messages passed to it with x-delay in the
header will be delayed by that many seconds.

An initial, but configurable, implementation of retrylists being delayed
upon each successive retry is also added, with a default progression of
delay times calculated from the number of retries on the retrylist.

Minimally tested so far, needs further integration testing.