Add Manual Acks and Persistence to All Messages
A manual acknowledgement is added to the end of all consumer callbacks.
This also includes a refactoring of error handling logic into a
wrapped_callback function to simplify and unfiy consumer callback
functionality.

Auto-acks have been removed, and persistence has been added to all
messages which will hopefully make the whole system more robust to
failures and downtime. It might also be worth making queues Durable in a
future commit.