Address StreamLostError Bug
This commit addresses issue #26, the StreamLostError/BrokenPipeError
which occurs in the server-api when attempting to send a message with a
stale connection. Behaviour is now to catch this error upon message
sending and attempt a single retry with a manually reset connection - as
per examples in the pika docs. This is now robust enough to run as
production - purportedly.

Also refactored the log message as it became clear that this behaviour
may need to be logged via the standard NLDS route i.e. through the
logging consumer. Also renamed self.queue to self.name in the consumer
to have a standard member variable across publisher and consumer, and
queue was essentially a name anyhow.