Implement Get Transfer Processor for Transact-ID
The get transfer processor is implemented for the simplest,
non-catalogue case whereby a user can request a particular transaction
id and any files contained therein. This will probably change when the
catalogue is eventually added as the workflow will probably be changed
to include getting file metadata from the catalogue first, whereas as of
this commit a 'stupid' workflow is implemented where the object store is
queried directly. The common nlds design feature of complete, retry and
fail lists is implemented, as are a number of other consumer config
variables.

Additional bugfixes for the put_transfer processor are also included.