Add a Remote Procedure Call client
This is a child-class of RabbitMQPublisher, which utilises an anonymous
exclusive queue, the reply_to and correlation_id headers, and a
correlation_id filter on callback to implement a Remote Procedure Call -
as per the instructions in rabbit tutorial 6. This in essence creates a
private consumer loop and ensures that only messages sent from the
client and then bounced from a specific consumer (sending RPC messages)
are processed.

This is a prototype but will be used for getting database queries back
to the user following a list, find, or stat command.