Fix monitoring state transition tracking
The state transition tracking for monitoring is fixed, with a retry flag
added to messages which are retrying instead of the state being set to
'RETRYING'. This is better logic to code and more informative to the
user.

Also bugfixed an issue with the publish_message method, whereby retry
and failed lists passed from the indexer were not being handled
properly.