Update _catalog_get() With New Aggregation Schema
Updated the get workflow to appropriately set up the archive_get
workflow process in the event that a file is requested that isn't on
Object Store. The file details generated now have to include both
tenancy and tape_url information, and the buckets are at that point
rejigged to align along aggregation boundaries instead of their original
transactions.