Add Migration for Change of State Values
This updates the database with the new state values needed for archive.
This will mostly affect complete and failed jobs, and for SQLite doesn't
actually change anything in the database as only the name is stored.
Haven't tested with PostgreSQL which may be more tricky as an ENUM
database variable is created and may not be able to be edited.