Change how tars are unpacked back into objectstore
The method for unpacking tars is fundamentally altered (and improved) to
upload files back to their original location. This required basiaclly a
complete rewrite of the archive-get callback (along with lots of
refactoring) as the way retries were being handled needed to completely
change. This also added a few extra methods to the catalog and extended
the catalog-get function slightly.

The new method for unpacking means that we will only attempt a retry for
archive-get with files that were in the original request (i.e. in the
transaction) and only in the case of a stream-time error. Any other
errors will be failed for rollback, as they largely only occur when
something is wrong with immutable information (the tar file, the message
info) and so successive retries would just be time consuming with no
chance of success. A possible exception is an xrootd error, which is a
subject for a future update.