Implement robust error handling for Archive-Put
Proper error handling for stream failures and mismatched checksums is
implemented on the archive-put consumer, so that the corrupted write is
removed from the disk cache before being written to tape. Incremental
naming is also implemented alongside, such that if any file is still in
the cache upon retry it does not clash (such as in the case of a broken
connection). This required downstream updates to the catalog to ensure
that all future reads from tape reference the correct, updated tar name.