Fix miscellaneous bugs with archive-get
Mutliple bugs and issues with the archive-get consumer callback are
fixed. This includes general improvements in logging quality, checking
for an empty prepare list before attempting to prepare, fixing an issue
the prepare logic for re-queueing a message, and probably others that i
am forgetting. This is now 'working' to a minimum extent, will require
further testing and tweaking though.