Fix phantom exception bug with successful retrieval
Previous code was throwing an exception during the stream step that was
being ignored by a mischevious continue in a finally block. This skips
the throwing of the exception, so files would write to object store
successfully and not be added to the completelist. This is now fixed,
and now additionally comes with additional logging of unexpected
exceptions so as not to happen again.

Also worth noting that objectstore writes do NOT require the connection
to be given up for reuse, no connection is passed back upon successful
write and this is all handled by minio.

A secondary change is included here where a specific delay can be fed to
the send_pathlist method, along with a new configuration option for the
archive-get, which is used to delay repeated checks of the prepare
request once it has been made.